+++ kernel/include/paging.h
@@ -9,6 +9,6 @@
 #include "multiboot.h"
 
 uint32_t initialize_paging();
-size_t calculate_max_memory(multiboot_info_t *mbd);
+uint32_t calculate_max_memory(multiboot_info_t *mbd);
 
 #endif
